Experimental: wip code to try adding a backend system for image I/O.
